+++ Sources/MapboxSearch/InternalAPI/Telemetry/EventsServiceProtocol.swift
@@ -5,4 +5,4 @@ protocol EventsServiceProtocol {
     func sendEvent(for event: Event, callback: EventsServiceResponseCallback?)
 }
 
-extension EventsService: EventsServiceProtocol { }
+extension EventsService: EventsServiceProtocol {}