+++ Sources/MapboxSearch/PublicAPI/Search Results/SearchResultSuggestionImpl.swift
@@ -53,7 +53,7 @@ class SearchResultSuggestionImpl: SearchResultSuggestion, CoreResponseProvider {
 
         self.id = coreResult.id
         self.mapboxId = coreResult.mapboxId
-        self.name = coreResult.names[0]
+        self.name = coreResult.names.first ?? ""
         self.address = coreResult.addresses?.first.map(Address.init)
         self.iconName = coreResult.icon
         self.serverIndex = coreResult.serverIndex?.intValue