+++ Sources/MapboxSearch/PublicAPI/SearchOptions.swift
@@ -37,7 +37,7 @@ public struct SearchOptions {
     /// Limit results to only those contained within the supplied bounding box.
     /// The bounding box cannot cross the 180th meridian.
     public var boundingBox: BoundingBox?
-    
+
     /// In case ``SearchOptions/boundingBox`` was applied, places search will look though all available tiles,
     /// ignoring the bounding box. Other search types (Address, POI, Category) will no be affected by this setting.
     /// In case ``SearchOptions/boundingBox`` was not applied - this param will not be used.
@@ -266,7 +266,8 @@ public struct SearchOptions {
             timeDeviation: timeDeviation,
             addonAPI: unsafeParameters,
             offlineSearchPlacesOutsideBbox: offlineSearchPlacesOutsideBbox,
-            ensureResultsPerCategory: nil, // TODO: Support multiple categories search and ability to ensure results per category.
+            ensureResultsPerCategory: nil,
+            // TODO: Support multiple categories search and ability to ensure results per category.
             attributeSets: nil
         )
     }