+++ Tests/MapboxSearchTests/Common/Extensions/LotLan+Extensions.swift
@@ -9,4 +9,4 @@ extension CoreBoundingBox {
             return false
         }
     }
-}   
+}